requirements.txt
autopep8==1.2
Flask==0.10.1
Flask-Migrate==1.4.0
Flask-RESTful==0.3.4
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
marshmallow==1.2.6
pep8==1.6.2
psycopg2==2.6
PyJWT==1.4.0
uWSGI==2.0.11.1


